refactor: move Identity class to a separate file
